Extend scores to allow multi lang, dupe people
Rather than the form /scores/:hole/:lang, the scores URLs are now
of the form /scores/:hole?/:lang1/:lang2.../show-duplicates?.
Where everything is optional! Sure makes the routing logic fun ;-)

The show duplicates flag at the end allows you to see every entry
for a partitcular hole and not just the best per person, it's not
the default and a later commit will probably removing the scoring
from this view, but it works, and is useful, thanks @AlexDaniel
for the suggestion :-)

The UI needs a lot of work since lang is still a dropdown despite
the backend accepting multiple values, but I wanted to get the code
merged while it kinda works.

Fixes #14