Fix score ordering on identical scores
The leaderboard was incorrectly ordered by score and not strokes.
In the event that two people score the same but with different
number of strokes (usually within one) it was possible for the
player with the fewer strokes to be placed after if their scores
matched.

I didn't notice it in the past because I had too few users :-P

Fixes #42