Prevent implict echo in "J"
This was making the quine far too easy so let's keep it sporting.
The secret to this seems to be to pass the code via a file rather
than STDIN, to accomplish this, add a crude C runner inbetween.
This can always be improved later, but for now it seems to work.

Updates #48