Fix for code mirror sizing issues on Firefox that prevent using the site. Fixes #121.
Prior to the next-gen merge, the creation of the code mirror happened on load in holes.js. Defer the creation of the code mirror to the onload callback. I moved the minimum amount of code to the callback that was necessary to for it to work well in Firefox. I didn't notice any negative consequences.