Update run-lang to bind-mount /dev
`rustc` requires access to `/dev/urandom` to run, but it was not available in
the rootfs. This change to run-lang.asm bind-mounts "rootfs/dev" to existing
"/dev" so `rustc` runs happily.

Also use certain env variables that `rustc` expects in the bash wrapper, in
particular `RUSTUP_HOME` is needed to find the installed Rust toolchain, and
`PATH` is required to find `cc` which is used by `rustc` for linking.

`RUST_BACKTRACE` is used for debugging `rustc` invocations as well as debugging
the user's program, but can be removed if that seems like "cheating".