Add rust Dockerfile + add to build-langs
Includes simple wrapper script that is staged to `/usr/bin/rust` which can
compile from stdin or a file and immediately executes the program.

Also add Rust to the build, and do a little deduplication of the list of
languages being built. Strip out "rustc " prefix to get Rust version.