Add support for Rust lang (#133)
* Make bash scripts a little more portable

build-assets:

  `stat` may have different arguments on different platforms, e.g. Linux usually
  accepts `-c %Y` to get mtime, but macOS and BSD may not.

  Instead, use the builtin test from `bash` to see whether `routes/assets.go` is
  older than the newest file in the assets dir.

build-langs:

  This script uses associate arrays, which are only in bash >= 4. Add a simple
  check at the beginning to ensure the minimum bash version is being used.

Makefile:

  Some versions of `find` do not accept predicates without a file path. This
  simply adds "." as the argument first

* Add rust Dockerfile + add to build-langs

Includes simple wrapper script that is staged to `/usr/bin/rust` which can
compile from stdin or a file and immediately executes the program.

Also add Rust to the build, and do a little deduplication of the list of
languages being built. Strip out "rustc " prefix to get Rust version.

* Add Rust to routes + views.

Add to SQL schema, versions, and types, and update comment in `runCode`.

* Add assets for Rust

CodeMirror Rust mode and "simple" mode taken from upstream:
https://github.com/codemirror/CodeMirror

Rust logo taken from official Rust artwork repo:
https://github.com/rust-lang/rust-artwork

* Update run-lang to bind-mount /dev

`rustc` requires access to `/dev/urandom` to run, but it was not available in
the rootfs. This change to run-lang.asm bind-mounts "rootfs/dev" to existing
"/dev" so `rustc` runs happily.

Also use certain env variables that `rustc` expects in the bash wrapper, in
particular `RUSTUP_HOME` is needed to find the installed Rust toolchain, and
`PATH` is required to find `cc` which is used by `rustc` for linking.

`RUST_BACKTRACE` is used for debugging `rustc` invocations as well as debugging
the user's program, but can be removed if that seems like "cheating".

* Use devtmpfs instead of bind mounting /dev

The parent container has the "devtmpfs" filesystem type, which allows us to
bind /dev in the same manner we bind /tmp, except it creates the expected
device files (/dev/null, /dev/urandom, etc.).

This also means that programs which want to use these files can (e.g. I was
able to write a Python program that reads from /dev/urandom).

* Only mount /dev/urandom instead of full devtmpfs

Instead of creating a full devtmpfs (which includes /dev/random, /dev/null, etc.) now only
/dev/urandom is mounted. This appears to be enough for `rustc` to work and probably exposes
less to the user's code.

* Merge remote-tracking branch 'upstream/master' into rust-lang

* Alpha order Raku > Python

* Update run-lang to mount /dev/random

We need to mount both /dev/urandom and /dev/random, since in some cases the
getrandom() syscall may not be available. This still avoids the use of devtmpfs
but works in cases where the kernel does not have sufficient entropy for
`getrandom()`.

* Remove duplicate Raku after merge conflict

* Fix copy-paste error in comment and merge problem