Make the matrix work on mobile
Switch for a table based layout that wasn't scalaing as more langs
were added to a CSS grid approach where langs can go across multiple
rows.

The old colours are back, but I'm not sold on them. The toggle feature
is new but I'm not sold on that either. Feedback welcome!

But I think it is important that this now works on mobile and will
scale as more languages are added.

Fixes #127
Updates #134