After submitting a shorter passing solution, update the solutions list and the picker tabs.
The picker tabs will show the length of the shorter solutions.
After switching languages and back, the shorter solution will be restored.
Additionally, show "..." instead of pass or fail while waiting for results. This rudimentary visual feedback helps users to see that the results have updated or that their program is still running.

Fixes #120.