Matrix: don't change language header color when hovering.
This works much better on iOS with Safari, because when you touch one of the headers, it immediately goes into its hover state and stays in that state until you touch a different one. The hover CSS rule made it difficult to see which languages were enabled.

This change also provides better visual feedback in general, because there's now immediate visual feedback on the headers after they are clicked.