hole.js: Fix an issue where runs can finish out of order.
* Only display the results for the latest run.
I recently had a case where the following sequence of events happened.
1. I typed in some code and hit ctrl+enter (run 1).
2. I realized that it was going to time out, typed a couple more characters to fix it, then hit ctrl+enter again (run 2).
3. Run 2 finished and displayed the pass status.
4. Several seconds later, run 1 finished and displayed the fail status.
5. The interface continued to display the fail status until I ran it again.

This change prevents items 4 and 5 from happening.
One easy way to test this is to use the following Python code: seconds=5;import time;time.sleep(seconds);print(seconds)
Start a run for a larger number of seconds, such as 5, then start another run for a lower number. Check the output value.

* Change status panel to use a grey background while waiting for results.
This makes it easier to see when the result is ready, since it will transition from grey to green/red.

* Remove previous change that showed "..." in place of Pass or Fail while waiting for results.
This isn't necessary with the background color change and it obfuscates the previous result.