Lucky tickets (#171)
* Add the lucky-tickets hole.

Note that the algorithm used in lucky-tickets.go is not the most efficient algorithm, but it is only used to generate the results for the random test cases and typically runs in less than 20 milliseconds. For the longest test case, it took about 2.5 seconds to generate the result on my machine, so I am including the precomputed result for performance. For many languages, a better algorithm will be required to solve the hole within the time constraint.

* hole/hole.go: Reduce complexity of the Play function.