Stop kicking me to Python
This is the minimal level of cleverness but I think it helps this
issue enough. We simply store the latest language you navigated to
anywhere in localstorage, and use this when you navigate to a hole.

This could be more clever later, but for now I think this is a win.

Fixes #165