Prototype a new /golfers/{golfer} route
This moves the old /prefs route to /settings and creates a new
user page with a more golfy name. The new routes has far nicer
looking trophies with emojis (fallback font still needed) and
splits the trophies and the matrix of holes across two distinct
pages for both clarity and load speed.

Obviously in the final version you'd only see the settings tab
when looking at your own profile while logged in.

Having the pecentage of people who won a given trophy show up is
a nice feature to have, the progress bar could be mistaken for
your own progress in that trophy but this design is similar to
how Steam does it. I would like to add partial progress for
trophies but that will need a complete datbase re-design so I'm
punting that for now. The number of people offers a good place
to put a link to a leaderboard of who won the trophy first so
I might add that easy feature soon.