More accurately count golfers on /stats
Trophies are the superset of solutions, it is possible to have a
trophy and not have a solution, the reverse is not possible.

This inflates the count by around 30 at the moment but that delta
might increase.

This was noticed while working on the new golfer page which shows
less than 100% of golfers having the trophy for submitting their
first solution.