Add a fallback emoji font
This replaces the old black and white one with a proper COLR/CPAL
colour emoji font. It it subset to only include used emojis and
then converted to woff2 to save space. CSS unicode ranges are
employed to prevent rendering on normal characters and local emoji
fonts are preferred to save bandwidth.