Optimise & simplify points view
tl;dr DISTINCT slow, GROUP BY fast(er)

Also drop the holes total as nothing was reading it. The average
query time to look up a single row by user_id went from 130ms to
100ms. Still way too slow, but it's a start.

This is part of a greater effort to speedup the /golfers/ route.
Next step might be to switch to a materialized view of points
that is updated when a better solution is submitted.

Even if the queries took the same amount of time the new one is
vastly simpler IMHO.