Move view JS into external files
They're included inline but minified and in such a way that they
shouldn't block document parsing.

The long term goal is to kill the currently cobbled together asset
pipeline. The only large script, codemirror will be loaded as a JS
module and will have a fingerprinted URL. Everything else will be
in the views dir and included inline. This is a trade-off that
makes developement and deployment easier. All CSS will be served
inline but only what is needed for that page.