hole.js: Fixed an issue where the user could be repeatedly prompted to restore a local solution.
Before this change, if you are logged in and submit a solution and then delete a character from it, then each time you reload the page, it would prompt you to restore the solution from local storage. This is because the version of the solution with the deleted character would remain in local storage. After a recent change (designed to minimize unnecessary prompting, ironically), the solution wouldn't go to local storage, because it matched, so the version with the deleted character would remain. The fix is to remove the local storage item if it matches the server's version of the solution.