Add “COBOL” language via GnuCOBOL + TCC.
#176

The free source format from Cobol 2002 is used and I updated the CodeMirror file to support this.

Docker image size is 6.12MB.

For the SVG, I traced the letters CBL, a common file extension for Cobol, from the cover of the COBOL 60 report to CODASYL (April 1960).

Also, modify Fortran version string to include the compiler name.