Auotmate the golfer sponsor flag
Right, rule of three, time to automate this. Thanks @rauchg,
@tomtheisen, @Andre-Lameirinhas.

The original GraphQL abstraction code has reached breaking point
as this query is on a viewer/user whereas all the other queries
to date were on the repository so it doesn't quite fit.

To combat this I've introduced the github.com/shurcooL/githubv4
library with the intent to port the other queries to this library
too.

Hopefully what will be left will be fewer and more readable lines
of code :-)