Generalise alert/info markup
Now we have a few different sources of alerts and information we
will generalise the CSS and markup.

This also employs a clever trick to extend the header to the full
page width so we apply the max-width directly to the body. This
simplifies the logic. Long term all pages should be consistent in
width.