Port some more icons to bootstrap
The remaining ones are harder and probably need more time to mature
or exist upstream first.

The most major difference is the use of a trophy for the scores
page. While bootstrap did have an ordered list, it had too much
whitespace so didn't look right. The overload with trophies is a
bit annoying, probably a good reason to go ahead with the rename
to achievements to be honest, it's what I as an Xbox player call
them anyways :-P

One nice aspect of this change is pretty much every replacement is
smaller than it was before which can only do good things for page
performance :-) I think the reason for this win is the viewbox is
only 16x16 in bootstrap while its a whopping 512x512 in font awe.