Fix an issue in the beta where the solution picker should be hidden.
Fix an issue where the solution picker could become visible when a user enters a longer solution.
The intention for the code that controls solution picker visibility was to allow logged-in users to switch to an auto-saved solution they entered when they were logged out. This change makes it more clear that the conditions only apply for autosaves for the other solution.