Add search/replace addons to codemirror (#300)
Also adds jump-to-line which may not really be needed considering the generally
short text contents of the golf editor, but gives a nice shortcut for it.

For vim mode, this seems to enable:
* Incremental highlighted search (/, ?, #, *, g#, g*)
* `:%s`, `:substitute`
* Probably some other stuff I don't know about

For non-vim mode this enables the shortcuts listed here:
https://codemirror.net/demo/search.html

Although some of these seem to conflict with browser shortcuts so functionality
may be limited that way depending on browser/OS (e.g. cmd+shift+R refreshes the
page on Firefox/macOS rather than opening the replace dialog).