Undo lucky tickets change
This is what happens when you blindly follow a linter without
thinking it through. Thanks for spotting @JayXon++

I think the linter isn't counting for there being a copy between
then make and the append, it's trying to catch cases when you
make with a len, then append, and have leading zero values.

Either way for now I'll just turn this linter off, it was new
anyways.