add language Lolcode
First pass at adding the Lolcode language. It's not well specced and a
little out of date, and probably does not have enough features to be
useful. I think the Dockerfile for the language needs some work to get
it to be smaller. Also the fact that lci is printing its version to
STDERR could be reworked.

lolcode.svg by John Gardner
(https://github.com/file-icons/icons/blob/master/svg/LOLCODE.svg)