Merge code table into solutions
This was needed in order to support the different scoring rules
around assembly. Assembly bytes are not the source code bytes so
the column is no longer a generated column, ATM it's still
computed the old way but this will change.

This new schema is less space effecient but sure is simpler, and
seems to be faster to boot!