Manage JS with esbuild
This commit switches to external, minified, cached JS files.
Each file can import multiple packages which will be bundled
by esbuild with the option of using TS/JSX in the future.

This commit does away with the need to run make editor or
commit a semi compiled version of CodeMirror 6 to the repo.

Closes #423.