Increase e2e reliability
On my machine, the DB takes longer to start up than the website,
this is probably because I develop against a dump of the live DB
which naturally contains a lot of rows. This causes the first one
or two test files to fail in e2e from time to time. Blocking on
the DB fixes this, let's see how GitHub CI handles it.