Update language dockerfiles to be easier to run.
Split up the ENTRYPOINT, so that ENTRYPOINT is the program to run the code and CMD is the default parameters, to obtain the version number.
When you run with no arguments, it prints the version number.
When you provide arguments, they replace the default arguments to get the version number.
Users can now use the containers like this:
docker run --rm -i codegolf/lang-ruby - < golf1.rb

Instead of needing to override the entrypoint:
docker run --rm -i --entrypoint /usr/bin/ruby codegolf/lang-ruby - < golf1.rb

For PowerShell, I had to add an - argument to get it to suppress the default --version argument when running the container from the command line.