Revert a recent change to hole.js, because it was incorrect.
The recent change made the solution picker appear sometimes when it shouldn't. I have modified the tests to detect this.

After careful consideration, I have decided that the behavior before that change was actually correct and have modified the tests to explicitly require that behavior.

I also made some other improvements to the tests. For example, in tests 2 and 3 in hole-logged-in.t and 1-3 in hole-logged-out.t, the results of checking the byte count were misleading, because I hadn't switched back to Raku after clearing localStorage.