Replace the prompt for restoring local solutions with a "Restore solution" link. (#464)
* Replace the prompt for restoring local solutions with a "Restore solution" link.

If a successful solution has been submitted, or loaded from the database, then the restore solution link appears whenever the contents of the code editor differs from that solution. Otherwise, it is hidden.

The UX for the prompt was not very good. Users had to decide right away whether to restore the local solution, so they usually said yes. If they then realized they didn't need it, they had to reload the page again and say no. The restore solution link makes it one click to restore the working solution, instead of having to reload the page and decline the prompt to restore the local solution.

Updated the e2e tests.

Fixes #306.

* Consolidate test code and remove a sleep call left in by mistake.

* Consolidate more test code. Reduce duplication of code.

For HoleWebDriver methods, make the context parameter required.

* Update a comment.

* hole.js: Move some code into functions to make it easier to read.

* Added two tests.