Add test case to brainfuck.go (#509)
* Add test case to brainfuck.go

A few messages on discord suggested there was a cheat on this hole so I'm adding a test case. The problem description and these test cases are not consistent. Maybe there should be an update to the description?

* Update hole/brainfuck.go

from discord convo. Canissimia, our brainfuck expert, suggested this kind of tail operation (and no-op) would never be in a real program and should be omitted

* Update hole/brainfuck.go

* Update hole/brainfuck.go