diff: Fix the e2e tests. Fixes #558.
The at() method for Array is very new and it isn't supported by the version of Firefox currently used by the e2e tests. It was implemented in Firefox around July 2021 and it's not supported by Safari on iOS. It's a bit to new to use in the code. It's unclear why esbuild didn't do something about this.