Minor improvements to sudoku
Drop the comment about DLX, I tried it and while it made the generation
faster and therefore we could clear more cells, that led to most
solutions timing out. So instead prefer the simplicity of a bruce force
generator too. Medium term I would like to merge the generator and the
solver as they're practically the same code.