Add a "Copy as JSON" button to some holes
Insipred by and closes #607, thanks @MichalMarsalek++

Adds structured JSON to some of the holes in holes.toml, using a
JSON string rather than a TOML structure for two reasons. One to
preserve order/formatting and two, TOML doesn't allow spaces in its
inline map syntax which makes the whole lot unreadable.

Where possible we use this data to build the preamble which makes this
patch actually net negative in line count!

The UI should probably communicate success on click but this is a good
first pass.