Fix make test
middleware is now loaded before routes so we don't do our hacky chdir to
fix the pwd in order to list some directories. Instead of doing that
let's make ls() handle the directory not being there and just return an
empty map, this should also make running the dev stack with no assets
more graceful.

We could probably also switch to things like embed to avoid the chdir
but that would mean restructuring the code so that the go is at the same
level (or higher) than the static files since you can't embed up a dir.