hole.js: simplify & inline populateLanguagePicker()
Migrate the building to JSX and tweak our factory to omit undefined
attributes and values like React would.

Hoist the sorting of the languages to a one time constant.

Also remove the superfluous IDs on tabs and instead use the link text in
the e2e like a human would.