hole.js: Inline & simplify populateSolutionPicker()
As part of this also fix a visual bug whereby when the picker is hidden
you still had a grid gap on both sides of it so effectively the editor
was too far down by one gap. The fix is to not empty the picker div but
to hide it instead. This required a small tweak to the e2e test where we
need to ensure the link is visible too.