Hide delete solution button for logged out users
Also introduce a new .hide class for easy hiding of elements. Use this
in a couple of places.

The ergonomics are a little suss since you have to invert the condition
but it makes up for that with not needing to know if the element should
be block, flex, grid, etc.

Also fold a few more selectors into their callsites, sure it's slightly
slower to lookup the element each time but you won't notice it and it
makes the code a lot more readable.