Add ranking view selectors to the hole page
This is inspired by the Xbox model where you can see overall, where you
fall, and where you and your friends fall.

Introduce a new API method to power it, it still needs to be added to
the OpenAPI docs.

This should also fix bugs around the "me" view when there aren't enough
other golfers on each side of you by moving the logic to Go, it's still
way too verbose but I'm struggling to simplify it.