Add hexagony.net link to hole
This commit ports the third party integration functionality from NG to
classic.

Prefer a non-target link on NG and classic as per our other external
links, this seems to be best practice as it allows users to opt-in to
new tab/window functionality if they so wish.

This commit also ported the status box to grid, this should've been done
years ago but it turned out a little non-trivial wrt overflow and those
vertical labels. Switch to a more modern way of making text vertical,
writing-mode vs transform, as a side effect the text is now 180° rotated
compared to before but still looks fine imo.