Bring NG up to feature parity
This is basically a port of classic to NG, the editor swapped out for
CM6 and a few NG specific things like PBM sprinkled on top.

Run like this for a while and then we can copy it over to classic.

As part of this we've de-duped the controllers as they really ought to
be doing no different work.