Run Perl solutions under -E
ATM we customise our perl to have say() always available, this is hacky
and I want to stop doing this. The obvious replacement is the -E switch
which enables all the features for that Perl version.

While we're here bump to the latest RC, it'll be out soon and I actually
care about Perl so we can use our corpus as some light testing for them.

This change isn't without fallout though, the main thing is when
signatures are enabled (as they are under -E), prototypes have to move
to a subroutine attribute which is probably too costly for most golf.

But the fallout is <100 solutions and simply enabling -E is both easier
to explain to people, and allows for all the new features and syntax in
golfing.