Remove Fortran todo
This is probably the same issue as K & Julia. Remove this for now and if
it comes back we need to a) pick a target arch that works for live and
see if we can pick which CPUs GH Actions run on and if now target the
lowest common denominator.