Move language notes from the blue info boxes to sample code. (#762)
Also removed info box content that was covered by the compiler version numbers and links on the about page.

For Assembly, I left the content in the info box, because moving the URLs to the sample code greatly increased the width of the sample code and affected the positions of the machine language annotations. I fixed a bug where the assembly language info content might be rendered briefly right after the page loads.