Fix banner spacing on non tabbed hole pages
This div only ought to be present on the tabbed hole page where it
becomes an x pinned to the top right of the banner. On every other page
it adds annoying spacing to the bottom of the banner.

Long term banners will be dismissable but for now keep this hack to only
the tabbed hole page.