Switch to sqlx
Longer term this should allow us to cut down on a lot of the boilerplate
when dealing with SQL.

Move some of the DB logic into a new package. Move the SQL files to an
sql dir instead.

This commit doesn't actually use sqlx, that'll happen slowly in
following commits which will more than make up for the lines added here.