Prevent trivial TeX quines
Requires TeX quines to have at least one backslash. This works
because any nontrival quine must use a macro, which must
start with backslash. In the other direction, the first backslash
in a file always starts a macro and never represents a verbatim
backslash.

This is entirely in the play.go logic. An explanation of the
backslash requirement is provided by stderr if violated, avoiding
the clutter of an info banner.