Fix another state exploit
Recently we updated the kernel the site runs on to remove
/proc/sched_debug as you could inject information into it by running
attempted solutions at the same time.

This time someone has found out how to do the same with
/proc/sys/fs/aio-nr.

/proc is a massive risk from my PoV but we need it for /proc/self/fd/0
to avoid writing code to disk for some languages.

Instead let's play whack-a-mole and just mount a dummy tmpfs over the
top of /proc/sys for now and maybe other paths in the future. This is
the lowest risk hack I could think of.