Fix Elixir sometimes failing e2e
Elixir walks /dev/fd on exit closing each handle. Prior to this commit
we lacked that symlink which led to Elixir sometimes hanging on exit and
having to be killed by the Go timeout. Why this seems to happen more
often on my laptop than production I can't say but the symlink is
harmless enough so we may as well have it.

This allowed a revert of the 10s e2e timeout increase which didn't work
anyways.

Finally make use of the new WaitDelay field when executing a command
from Go, this sets the maximum time Go should wait until killing a hung
process. This should lead to fewer zombies and just makes sense.