Refactor solution saving
The new logic is longer, but far clearer imo. Instead of trying to stuff
it all into one upsert, split the logic into a select with conditional
inserts or updates. This would be slower if we had the latency of going
over to the DB each time but because this is PL/pgSQL all the execution
happens on the DB. Finally wrap the logic in a for loop over scoring to
make it a bit more DRY.

This makes it a lot easier to add two highly wanted features to this
logic, we can now trivally update a history table when the solution is
improved and we can fairly easily add support for saving timings to the
DB too with their own complicated conditional update logic.

Longer term the rest of the logic around computing the ranking
improvement ought to fold into the for loop but that's fairly
non-trivial with the way the return variables are named atm, we probably
want to return some of kind of array/struct or perhaps even JSON.

Updates #592
Updates #869