Use a custom font to fix FEN alignment
Thanks to @btnlq for the font work.

I've tried to limit the scope of the font to just the FEN hole as we
really don't want to mess with popular characters like space everywhere
as it would probably ruin monospace on a lot of fonts. I suspect longer
term we probably want to combine this chess font (Segoe UI Symbol) and
our Source Code Pro font into one monospace font for all of code.golf.

Updates #645