Redesign how we store things in context
Rather than store values directly in the request context, store a single
struct pointer as early as possible and then store various things inside
that.

This allows us to fix a long standing issue where middleware above can't
access values below and as such things like error pages don't have the
golfer in the top right or respect the dark/light preferences of the
golfer.

Add a small new feature where we log the golfer's name in the logs now
that we can get it for free.

The getters could probably be refactored further, perhaps adding small
until methods directly to the routes package and then folding the
session package directly into the middleware package but for now this is
a good place to leave it.

Fixes #842