Extract golfer link HTML into a template
Do this by moving the country flag from something that is looked up in
the template from a global to something which can be scanned directly
from an ID in the DB.

A lot of the churn is from a centralised template requiring consistent
naming wrt Login vs Name.

The Scanner interface is powerful, future commits should extend it to
Hole and Lang.