The great Run refactor
Primarily this commit renames Scorecard to Run (a MUCH better name) and
passes all runs down to the controller. We still return the legacy JSON
with a single run to avoid breaking the current UI and any consumers for
now but we also now have an additional "runs" field that contains the
results of every run which future commits will display in the UI.

This refactoring was also an excuse to add correctly lowercased JSON
names for all the new fields.

Updates #895