Introduce PHP 8
I've been putting this off for a while but like Python 3 we eventually
have to move with the times.

Introduce PHP 8 as just PHP and rename the current PHP to PHP 7. This
legacy language will stick around for a bit but I don't think forever.
Copy all existing solutions across so they're saved against both
languages with the original timestamps but re-testing causes hundreds of
them to fail as PHP 8 is more strict around barewords.