Finally bump Go
The change in 1.20 to stop pre-compiling libraries really did a number
on us. I've switched us to a simpler `go run` workflow and a shell
script for now. This change opens us up to trivial quines but it was
needed to make sense of what was breaking, don't make perfect the enemy
of working!