HTML escape top-level Stderr, fixes e2e
This is a regression from when we started HTML escaping the multi run
output, we probably ought to be using pointers so less copying occurs
and order doesn't matter so much but tbh the real fix is to drop the
legacy top level output.

So for now, do this to fix e2e and add a small optimisation to avoid an
extra copy.