Unify navigation HTML generation.
There is now a common template for handling sidebar on desktop and
dropdown on mobile navigation generation from a set of structs.

For most pages we can define these at startup from the holes/langs TOML
files but for wiki we query the database in the controller. To that end
the wiki controller has pretty much been re-written to be simpler.

The recent-holes template and CSS has been spat out to make the
navigation possible, that's more repetition than I'd like but we're
still a net reduction in LoC.

Now that navigation is handled generically we don't need to pass so many
ad-hoc params down from the controllers, I've tidied recent solutions
but more can definitely follow.

This commit is pretty rough around the edges, especially in naming but
it's a step in the right direction and unifies a lot of code.