Link to rankings from single solution page
Ideally these would be deep links, like ?page=2#56 or something but we'd
need row_number in the rankings table for that.

Remove failing for now as it doesn't serve much purpose since we only
link to non-failing solutions and only non-failing solutions have ranks.

Finally link to this page from the golfer's matrix now that it's one
small hop to rankings, this way history is far more discoverable.