Validate golfer usernames
This doesn't really do much as we currently get all of our usernames
from GitHub but in the near future we will allow golfers to change their
username. This regex is a superset of what GitHub allows, ours is longer
and doesn't care about having leading or trailing hyphens. 42 just
seemed like a better number ;-)