Add static test cases for Mahjong. (#1047)
When checking for a standard hand with four melds and a pair, it's possible to cheat a little bit by assuming we have one extra of one of the the 34 tiles. If we had a pair, it would become a triple, which is ok. But if we didn't have a pair, but the remaining two tiles that aren't part of a meld could be used as part of a sequence, then we could incorrectly detect the hand as complete by treating these two tiles as a pair. This often works, because we don't always have random test cases that fit this pattern. This adds two static test cases that fit that pattern.