Remove Clojure implicit output of the last form. Fixes #1060.
Clojure currently outputs the result of the last form, unless it is nil. This doesn't appear to be a standard feature of the language. Instead, it seems to be a quirk of the Babashka interpreter that only occurs when providing code via a command line argument. It's probably best to disable it, since people won't be expecting it and it's generally not too useful.