Update discord bot to announce the user name of the previous record holder.
If more than one user held the previous record, the number of users is shown instead.
When saving a new solution, retrieve more info from the db about the previous best score.
If the scoring was previously a diamond, retrieve the id of the golfer who held that diamond. Also return the number of golfers that were tied for gold.