Update the discord bot to log matched diamonds #629
Update existing discord messages so that the golfer who achieved a new diamond isn't included in the number/name of the golfer who held the previous record.
If user A and B are tied for gold and then A achieves a new diamond, the discord post will now show (A) instead of (2 golfers) following the old score.

Remove unused fields beat_bytes and beat_chars from the save_solution SQL function.

Fixed a bug related to the client side notifications where we weren't showing the new diamond notification, if you previously had gold.