Fix a bug where the discordbot could show the wrong score for a new diamond. (#1071)
I forgot to filter out failing solutions when checking for the shortest solution for a hole. I copied this code from elsewhere and it would have previously had a bug where it posted the wrong old stroke count for a new diamond, but it would have been less noticeable, because the number was guaranteed to be greater than the strokes for the new diamond.