Updates to the JS score popups.
Users always see their current score in points, in addition to the offset, because the mini-scoreboard doesn't show this.
If their rank has changed, their old and new rank is shown.
When the user submits their first solution for a hole/lang, they will see their rank and score in points in the popups.