Correctly wrap long links on a golfer's wall
Do this by dropping them being flex, even inline-flex didn't seem to
work. One casualty of this is we no longer have profile images when a
golfer follows you but that's not a huge loss. Cheevo "images" still
show as they're emoji and flow into the text properly, but they had to
be moved outside of the link to avoid having ugly underlines on hover.

All in all this is a change full of compromises but I really hated how
bad the lack of correct wrapping looked on mobile so this is a win!