🥧 This commit ate all the pies!
Redesign the stats page to follow the same navigation style as other
pages, split the tables off into other routes, and drop the pie charts.
This allows for future commits to add more statistics easily.

The pie charts were some of the oldest code in the site and didn't look
that good.

Future commits will add graphs of things like golfers over time, because
a) graphs are cool, b) I'd like to see that growth line, and c) so that
the stats icon in the header still makes sense now that the pie charts
are gone.